+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog for package grepros
 * refactor database and SQL sinks onto a common base
 * allow specifying dialect options in Postgres/SQLite output
 * allow overriding table and view names in SQL dialects
+* allow specifying field value adapters in SQL dialects
 * drop meta-table from Postgres output
 * drop default value from --publish-prefix option
 * fix error in parsing subtypes from message definitions